+++ e2e/resolve-conditions/deno-env.js
@@ -3,12 +3,14 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- *
- * @jest-environment jest-environment-node
  */
 
-import {fn} from 'fake-dual-dep';
+'use strict';
 
-test('returns correct message', () => {
-  expect(fn()).toEqual('hello from node');
-});
+const NodeEnv = require('jest-environment-node');
+
+module.exports = class DenoEnvWithConditions extends NodeEnv {
+  exportConditions() {
+    return ['deno'];
+  }
+};